refactor: added scaling_factor and power_factor to chunk_frame_generator
Via the newly introduced arguments it is possible to estimate the allowed window size based on the complexity of the problem one would like to implement.

For example: If a user needs all combinations of two trajectories in the virtual memory - then the best solution would be to choose a power_factor of 2.